feat: Реализация авторизации

+++ controllers/cards.js
@@ -1,54 +1,44 @@
 const Card = require('../models/card');
-const { NOT_FOUND_ERROR, DEFAULT_ERROR, DATA__ERROR } = require('../utils/constants');
+const DefaultError = require('../errors/DefaultError');
+const handleErrors = require('../errors/handleError');
 
-const getAllCards = (req, res) => {
+const getAllCards = (req, res, next) => {
   Card.find({})
     .then((cards) => {
       res.send({ data: cards });
     })
-    .catch(() => res.status(DEFAULT_ERROR).send({ message: 'Что-то пошло не так' }));
+    .catch(() => next(new DefaultError('Что-то пошло не так')));
 };
 
-const createCard = (req, res) => {
+const createCard = (req, res, next) => {
   const { name, link } = req.body;
   Card.create({ name, link, owner: req.user._id })
     .then((card) => {
       res.send({ data: card });
     })
     .catch((err) => {
-      if (err.name === 'ValidationError') {
-        const message = Object.values(err.errors)
-          .map((error) => error.message)
-          .join('; ');
-        res.status(DATA__ERROR).send({ message });
-      } else {
-        res.status(DEFAULT_ERROR).send({ message: 'Что-то пошло не так' });
-      }
+      handleErrors(err, req, res, next);
     });
 };
 
-const deleteCard = (req, res) => {
+const deleteCard = (req, res, next) => {
   const { cardId } = req.params;
-  Card.findByIdAndRemove(cardId)
-    .orFail(() => {
-      throw new Error('Not found');
-    })
-    .then((card) => {
-      res.send({ data: card });
-    })
-    .catch((err) => {
-      if (err.message === 'Not found') {
-        res.status(NOT_FOUND_ERROR).send({ message: 'Карточка с указанным _id не найдена.' });
-      }
-      if (err.name === 'CastError') {
-        res.status(DATA__ERROR).send({ message: 'Передан некорректный id' });
-      } else {
-        res.status(DEFAULT_ERROR).send({ message: 'Что-то пошло не так' });
-      }
-    });
+  if (req.user._id === req.owner) {
+    Card.findByIdAndRemove(cardId)
+      .orFail(() => {
+        throw new Error('Not found');
+      })
+      .then((card) => {
+        res.send({ data: card });
+      })
+      .catch((err) => {
+        handleErrors(err, req, res, next);
+      });
+  }
+  throw new Error('Нет доступа');
 };
 
-const likeCard = (req, res) => {
+const likeCard = (req, res, next) => {
   const { cardId } = req.params;
   Card.findByIdAndUpdate(cardId, { $addToSet: { likes: req.user._id } }, { new: true })
     .orFail(() => {
@@ -58,18 +48,11 @@ const likeCard = (req, res) => {
       res.send({ data: card });
     })
     .catch((err) => {
-      if (err.message === 'Not found') {
-        res.status(NOT_FOUND_ERROR).send({ message: 'Передан несуществующий _id карточки.' });
-      }
-      if (err.name === 'CastError') {
-        res.status(DATA__ERROR).send({ message: 'Передан некорректный id' });
-      } else {
-        res.status(DEFAULT_ERROR).send({ message: 'Что-то пошло не так' });
-      }
+      handleErrors(err, req, res, next);
     });
 };
 
-const dislikeCard = (req, res) => {
+const dislikeCard = (req, res, next) => {
   const { cardId } = req.params;
   Card.findByIdAndUpdate(cardId, { $pull: { likes: req.user._id } }, { new: true })
     .orFail(() => {
@@ -79,14 +62,7 @@ const dislikeCard = (req, res) => {
       res.send({ data: card });
     })
     .catch((err) => {
-      if (err.message === 'Not found') {
-        res.status(NOT_FOUND_ERROR).send({ message: 'Карточка с указанным _id не найдена.' });
-      }
-      if (err.name === 'CastError') {
-        res.status(DATA__ERROR).send({ message: 'Передан некорректный id' });
-      } else {
-        res.status(DEFAULT_ERROR).send({ message: 'Что-то пошло не так' });
-      }
+      handleErrors(err, req, res, next);
     });
 };
